+++ flumine/flumine.py
@@ -16,7 +16,6 @@ class Flumine(BaseFlumine):
             while True:
                 event = self.handler_queue.get()
                 if event.EVENT_TYPE == EventType.TERMINATOR:
-                    self._process_end_flumine()
                     break
 
                 elif event.EVENT_TYPE == EventType.MARKET_CATALOGUE: